Rename project to avoid a conflict with nuget package
