+++ source/EnergySupplying.Contracts/AssemblyInfo.cs
@@ -12,7 +12,8 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-namespace Contracts.BusinessRequests.MoveIn
+namespace EnergySupplying.Contracts;
+
+public class AssemblyInfo
 {
-    public record Response(IEnumerable<string> ValidationErrors, string ProcessId = "");
 }
