+++ source/EnergySupplying.Contracts/BusinessRequests/MoveIn/Request.cs
@@ -12,8 +12,13 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-namespace Contracts;
-
-public class AssemblyInfo
+namespace EnergySupplying.Contracts.BusinessRequests.MoveIn
 {
+    public record RequestV2(
+        string? AccountingPointNumber,
+        string? EnergySupplierNumber,
+        string? EffectiveDate,
+        Customer? Customer);
+
+    public record Customer(string? Name, string? Number);
 }
