+++ source/EnergySupplying.Contracts/EnergySupplying.Contracts.csproj
