+++ source/Messaging.Infrastructure/Messaging.Infrastructure.csproj
@@ -39,13 +39,14 @@ limitations under the License.
   </ItemGroup>
 
     <ItemGroup>
-      <ProjectReference Include="..\Contracts\Contracts.csproj" />
+      <ProjectReference Include="..\EnergySupplying.Contracts\EnergySupplying.Contracts.csproj" />
       <ProjectReference Include="..\Messaging.CimMessageAdapter\Messaging.CimMessageAdapter.csproj" />
       <ProjectReference Include="..\Messaging.Application\Messaging.Application.csproj" />
     </ItemGroup>
 
     <ItemGroup>
       <Folder Include="Configuration\SystemTime" />
+      <Folder Include="Transactions\MoveIn\Contracts" />
     </ItemGroup>
 
     <ItemGroup>
