+++ source/Messaging.IntegrationTests/TestDoubles/HttpClientSpy.cs
@@ -17,7 +17,7 @@ using System.Collections.Generic;
 using System.Net;
 using System.Net.Http;
 using System.Threading.Tasks;
-using Contracts.BusinessRequests.MoveIn;
+using EnergySupplying.Contracts.BusinessRequests.MoveIn;
 using Messaging.Infrastructure.Transactions;
 using Newtonsoft.Json;
 using Newtonsoft.Json.Linq;
