+++ source/Processing.Application/Processing.Application.csproj
@@ -30,7 +30,7 @@ limitations under the License.
   </ItemGroup>
 
   <ItemGroup>
-    <ProjectReference Include="..\Contracts\Contracts.csproj" />
+    <ProjectReference Include="..\EnergySupplying.Contracts\EnergySupplying.Contracts.csproj" />
     <ProjectReference Include="..\Processing.Domain\Processing.Domain.csproj" />
   </ItemGroup>
 </Project>
